App/Backend/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
click==8.0.4
colorama==0.4.4
coloredlogs==15.0.1
Django==4.0.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-WTF==1.0.0
humanfriendly==10.0
image==1.5.33
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
numpy==1.22.4
opencv-python==4.6.0.66
Pillow==9.1.1
pycodestyle==2.8.0
pyreadline3==3.4.1
scipy==1.8.1
six==1.16.0
spicy==0.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
Werkzeug==2.0.3
WTForms==3.0.1
